requirements.txt
text
torch
torchvision
numpy
matplotlib
scikit-learn
streamlit
pillow
